+++ scripts/repo/winesapos-build-repo.sh
@@ -83,6 +83,7 @@ makepkg_fn linux-firmware-valve
 rm -f ${OUTPUT_DIR}/linux-firmware-neptune* ${OUTPUT_DIR}/steamdeck-dsp-*
 makepkg_fn nexusmods-app-bin
 makepkg_fn oxp-sensors-dkms-git
+makepkg_fn reiserfsprogs
 makepkg_fn reiserfs-defrag
 makepkg_fn ssdfs-tools
 makepkg_fn steamtinkerlaunch-git